make operator a package
