+++ netket/operator/local_values.py
@@ -1,237 +1,12 @@
-from ._C_netket.operator import *
-
-from ._C_netket.operator import (
+from .._C_netket.operator import (
     _local_values_kernel,
     _der_local_values_kernel,
     _der_local_values_notcentered_kernel,
-    _rotated_grad_kernel,
 )
-from ._C_netket.machine import DensityMatrix
+
+from .._C_netket.machine import DensityMatrix as _DensityMatrix
 
 import numpy as _np
-
-
-def Ising(hilbert, h, J=1.0):
-    """
-    Constructs a new ``Ising`` given a hilbert space, a transverse field,
-    and (if specified) a coupling constant.
-
-    Args:
-        hilbert: Hilbert space the operator acts on.
-        h: The strength of the transverse field.
-        J: The strength of the coupling. Default is 1.0.
-
-    Examples:
-        Constructs an ``Ising`` operator for a 1D system.
-
-        >>> import netket as nk
-        >>> g = nk.graph.Hypercube(length=20, n_dim=1, pbc=True)
-        >>> hi = nk.hilbert.Spin(s=0.5, graph=g)
-        >>> op = nk.operator.Ising(h=1.321, hilbert=hi, J=0.5)
-        >>> print(op.hilbert.size)
-        20
-    """
-    sigma_x = _np.array([[0, 1], [1, 0]])
-    sz_sz = _np.array([[1, 0, 0, 0], [0, -1, 0, 0], [0, 0, -1, 0], [0, 0, 0, 1]])
-    return GraphOperator(hilbert, siteops=[-h * sigma_x], bondops=[J * sz_sz])
-
-
-def Heisenberg(hilbert, J=1, sign_rule=None):
-    """
-    Constructs a new ``Heisenberg`` given a hilbert space.
-
-    Args:
-        hilbert: Hilbert space the operator acts on.
-        J: The strength of the coupling. Default is 1.
-        sign_rule: If enabled, Marshal's sign rule will be used. On a bipartite
-                   lattice, this corresponds to a basis change flipping the Sz direction
-                   at every odd site of the lattice. For non-bipartite lattices, the
-                   sign rule cannot be applied. Defaults to True if the lattice is
-                   bipartite, False otherwise.
-
-    Examples:
-     Constructs a ``Heisenberg`` operator for a 1D system.
-
-        >>> import netket as nk
-        >>> g = nk.graph.Hypercube(length=20, n_dim=1, pbc=True)
-        >>> hi = nk.hilbert.Spin(s=0.5, total_sz=0, graph=g)
-        >>> op = nk.operator.Heisenberg(hilbert=hi)
-        >>> print(op.hilbert.size)
-        20
-    """
-    if sign_rule is None:
-        sign_rule = hilbert.graph.is_bipartite
-
-    sz_sz = _np.array([[1, 0, 0, 0], [0, -1, 0, 0], [0, 0, -1, 0], [0, 0, 0, 1]])
-    exchange = _np.array([[0, 0, 0, 0], [0, 0, 2, 0], [0, 2, 0, 0], [0, 0, 0, 0]])
-    if sign_rule:
-        if not hilbert.graph.is_bipartite:
-            raise ValueError("sign_rule=True specified for a non-bipartite lattice")
-        heis_term = sz_sz - exchange
-    else:
-        heis_term = sz_sz + exchange
-    return GraphOperator(hilbert, bondops=[J * heis_term])
-
-
-def sigmax(hilbert, site):
-    """
-    Builds the sigma_x operator acting on the `site`-th of the Hilbert
-    space `hilbert`.
-
-    If `hilbert` is a non-Spin space of local dimension M, it is considered
-    as a (M-1)/2 - spin space.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-    S = (N - 1) / 2
-
-    D = [_np.sqrt((S + 1) * 2 * a - a * (a + 1)) for a in _np.arange(1, N)]
-    mat = _np.diag(D, 1) + _np.diag(D, -1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def sigmay(hilbert, site):
-    """
-    Builds the sigma_y operator acting on the `site`-th of the Hilbert
-    space `hilbert`.
-
-    If `hilbert` is a non-Spin space of local dimension M, it is considered
-    as a (M-1)/2 - spin space.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-    S = (N - 1) / 2
-
-    D = _np.array(
-        [1j * _np.sqrt((S + 1) * 2 * a - a * (a + 1)) for a in _np.arange(1, N)]
-    )
-    mat = _np.diag(D, -1) + _np.diag(-D, 1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def sigmaz(hilbert, site):
-    """
-    Builds the sigma_z operator acting on the `site`-th of the Hilbert
-    space `hilbert`.
-
-    If `hilbert` is a non-Spin space of local dimension M, it is considered
-    as a (M-1)/2 - spin space.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-    S = (N - 1) / 2
-
-    D = _np.array([2 * m for m in _np.arange(S, -(S + 1), -1)])
-    mat = _np.diag(D, 0)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def sigmam(hilbert, site):
-    """
-    Builds the $sigma_- = sigma_x - im * sigma_y$ operator acting on the
-    `site`-th of the Hilbert space `hilbert`.
-
-    If `hilbert` is a non-Spin space of local dimension M, it is considered
-    as a (M-1)/2 - spin space.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-    S = (N - 1) / 2
-
-    S2 = (S + 1) * S
-    D = _np.array([_np.sqrt(S2 - m * (m - 1)) for m in _np.arange(S, -S, -1)])
-    mat = _np.diag(D, -1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def sigmap(hilbert, site):
-    """
-    Builds the $sigma_- = sigma_x + im * sigma_y$ operator acting on the
-    `site`-th of the Hilbert space `hilbert`.
-
-    If `hilbert` is a non-Spin space of local dimension M, it is considered
-    as a (M-1)/2 - spin space.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-    S = (N - 1) / 2
-
-    S2 = (S + 1) * S
-    D = _np.array([_np.sqrt(S2 - m * (m + 1)) for m in _np.arange(S - 1, -(S + 1), -1)])
-    mat = _np.diag(D, 1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def destroy(hilbert, site):
-    """
-    Builds the boson destruction operator acting on the `site`-th of the
-     Hilbert space `hilbert`.
-
-    If `hilbert` is a non-Bosonic space of local dimension M, it is considered
-    as a bosonic space of local dimension M.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-
-    D = _np.array([_np.sqrt(m) for m in _np.arange(1, N)])
-    mat = _np.diag(D, 1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def create(hilbert, site):
-    """
-    Builds the boson creation operator acting on the `site`-th of the
-     Hilbert space `hilbert`.
-
-    If `hilbert` is a non-Bosonic space of local dimension M, it is considered
-    as a bosonic space of local dimension M.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-
-    D = _np.array([_np.sqrt(m) for m in _np.arange(1, N)])
-    mat = _np.diag(D, -1)
-    return LocalOperator(hilbert, mat, [site])
-
-
-def number(hilbert, site):
-    """
-    Builds the number operator acting on the `site`-th of the
-    Hilbert space `hilbert`.
-
-    If `hilbert` is a non-Bosonic space of local dimension M, it is considered
-    as a bosonic space of local dimension M.
-
-    :param hilbert: The hilbert space
-    :param site: the site on which this operator acts
-    :return: a nk.operator.LocalOperator
-    """
-    N = hilbert.local_size
-
-    D = _np.array([m for m in _np.arange(0, N)])
-    mat = _np.diag(D, 0)
-    return LocalOperator(hilbert, mat, [site])
 
 
 def _local_values_impl(op, machine, v, log_vals, out):
@@ -295,7 +70,7 @@ def local_values(op, machine, v, log_vals=None, out=None):
     """
 
     # True when this is the local_value of a densitymatrix times an operator (observable)
-    is_op_times_op = isinstance(machine, DensityMatrix) and not isinstance(
+    is_op_times_op = isinstance(machine, _DensityMatrix) and not isinstance(
         op, LocalLiouvillian
     )
