+++ setup.py
@@ -160,7 +160,7 @@ setup(
     url="http://github.com/netket/netket",
     author_email="netket@netket.org",
     license="Apache 2.0",
-    packages=["netket", "netket.machine", "netket.sampler"],
+    packages=["netket", "netket.machine", "netket.sampler", "netket.operator"],
     ext_modules=[CMakeExtension("netket._C_netket")],
     long_description="""NetKet is an open - source project delivering cutting - edge
          methods for the study of many - body quantum systems with artificial
